requirements.txt
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
docopt==0.6.2
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
pipreqs==0.4.11
portalocker==2.7.0
pycparser==2.21
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.13.0
